LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-not-executed-after-shutdown-expected.txt


PASS Script is not executed after script thread is shutdown
